move all client code into client subdir
