+++ kernel/Cargo.toml
@@ -39,7 +39,7 @@ arrow-ord = { version = "^49.0", optional = true }
 arrow-schema = { version = "^49.0", optional = true }
 futures = { version = "0.3", optional = true }
 object_store = { version = "^0.8.0", optional = true }
-# Used in default and simple client
+# Used in default and sync client
 parquet = { version = "^49.0", optional = true }
 
 # optionally used with default client (though not required)
@@ -47,7 +47,7 @@ tokio = { version = "1", optional = true, features = ["rt-multi-thread"] }
 
 [features]
 arrow-conversion = ["arrow-schema"]
-default = ["simple-client"]
+default = ["sync-client"]
 default-client = [
   "arrow-array",
   "arrow-conversion",
@@ -64,7 +64,7 @@ default-client = [
 ]
 
 developer-visibility = []
-simple-client = [
+sync-client = [
   "arrow-array",
   "arrow-conversion",
   "arrow-json",
