+++ kernel/src/client/default/expression.rs
@@ -20,7 +20,7 @@ use itertools::Itertools;
 use crate::error::{DeltaResult, Error};
 use crate::expressions::{BinaryOperator, Expression, Scalar, UnaryOperator, VariadicOperator};
 use crate::schema::{DataType, PrimitiveType, SchemaRef};
-use crate::simple_client::data::SimpleData;
+use crate::client::sync::data::SimpleData;
 use crate::{EngineData, ExpressionEvaluator, ExpressionHandler};
 
 // TODO leverage scalars / Datum
