+++ kernel/src/client/sync/get_data.rs
