+++ kernel/src/lib.rs
@@ -60,13 +60,10 @@ pub use table::Table;
 #[cfg(feature = "arrow-conversion")]
 pub mod arrow_conversion;
 
-#[cfg(feature = "simple-client")]
-pub mod simple_client;
-
-#[cfg(feature = "default-client")]
+#[cfg(any(feature = "default-client", feature = "sync-client"))]
 pub mod client;
 #[cfg(feature = "default-client")]
-pub use client::*;
+pub use client::default::*;
 
 /// Delta table version is 8 byte unsigned int
 pub type Version = u64;
