+++ kernel/src/table.rs
@@ -48,7 +48,7 @@ mod tests {
     use std::path::PathBuf;
 
     use super::*;
-    use crate::simple_client::SimpleClient;
+    use crate::client::sync::SimpleClient;
 
     #[test]
     fn test_table() {
